sibling_position method
Sibling position method is implemented. In order to keep the node struct size small, sibling position is not stored in the vector elements. Therefore, this method requires a linear search; complexity is linear in the number of children of the parent. This is an intentional choice for now which might change in the future if iterators require sibling index to be stored.